requirements.txt
google-cloud-bigquery
gunicorn
google-cloud-aiplatform
streamlit